backend/requirements.txt
fastapi
uvicorn
python-dotenv
python-multipart
pyjwt
httpx
sqlalchemy
passlib
python-jose
bcrypt
email-validator